Made domain-moving deactivateable, added realtime patch, added possibility to login multiple times
git-svn-id: file:///var/svn/trunk@2323 45fdb5c4-e40b-0410-b369-9aab4fe9a275

+++ syscp/index.php
@@ -116,7 +116,11 @@ if($action == 'login')
 				$language = $settings['panel']['standardlanguage'];
 			}
 
-			$db->query("DELETE FROM `" . TABLE_PANEL_SESSIONS . "` WHERE `userid` = '" . (int)$userinfo['userid'] . "' AND `adminsession` = '" . $db->escape($userinfo['adminsession']) . "'");
+			if($settings['session']['allow_multiple_login'] != '1')
+			{
+				$db->query("DELETE FROM `" . TABLE_PANEL_SESSIONS . "` WHERE `userid` = '" . (int)$userinfo['userid'] . "' AND `adminsession` = '" . $db->escape($userinfo['adminsession']) . "'");
+			}
+
 			$db->query("INSERT INTO `" . TABLE_PANEL_SESSIONS . "` (`hash`, `userid`, `ipaddress`, `useragent`, `lastactivity`, `language`, `adminsession`) VALUES ('" . $db->escape($s) . "', '" . (int)$userinfo['userid'] . "', '" . $db->escape($remote_addr) . "', '" . $db->escape($http_user_agent) . "', '" . time() . "', '" . $db->escape($language) . "', '" . $db->escape($userinfo['adminsession']) . "')");
 
 			if($userinfo['adminsession'] == '1')
